+++ src/main/java/com/cloud_idaas/core/implementation/authentication/jwt/StaticPrivateKeyAssertionProvider.java
@@ -73,13 +73,33 @@ public class StaticPrivateKeyAssertionProvider implements JwtClientAssertionProv
         jws.setPayload(claims.toJson());
         jws.setKey(privateKey);
 
-        if (privateKey instanceof ECPrivateKey) {
-            jws.setAlgorithmHeaderValue(AlgorithmIdentifiers.ECDSA_USING_P256_CURVE_AND_SHA256);
-        } else if (privateKey instanceof RSAPrivateKey) {
-            jws.setAlgorithmHeaderValue(AlgorithmIdentifiers.RSA_USING_SHA256);
+        final String algorithm;
+        if (privateKey instanceof RSAPrivateKey) {
+            int keySize = ((RSAPrivateKey) privateKey).getModulus().bitLength();
+            if (keySize <= 2048) {
+                algorithm = AlgorithmIdentifiers.RSA_USING_SHA256;
+            } else if (keySize <= 3072) {
+                algorithm = AlgorithmIdentifiers.RSA_USING_SHA384;
+            } else {
+                algorithm = AlgorithmIdentifiers.RSA_USING_SHA512;
+            }
+        } else if (privateKey instanceof ECPrivateKey) {
+            int orderBitLength = ((ECPrivateKey) privateKey).getParams().getOrder().bitLength();
+            if (orderBitLength <= 256) {
+                algorithm = AlgorithmIdentifiers.ECDSA_USING_P256_CURVE_AND_SHA256;
+            } else if (orderBitLength <= 384) {
+                algorithm = AlgorithmIdentifiers.ECDSA_USING_P384_CURVE_AND_SHA384;
+            } else if (orderBitLength <= 521) {
+                algorithm = AlgorithmIdentifiers.ECDSA_USING_P521_CURVE_AND_SHA512;
+            } else {
+                throw new CredentialException(ErrorCode.NOT_SUPPORTED_WEB_KEY.getCode(),
+                        "Unsupported EC curve with order bit length: " + orderBitLength);
+            }
         } else {
             throw new CredentialException(ErrorCode.NOT_SUPPORTED_WEB_KEY.getCode(), "Not supported web key: " + privateKey);
         }
+
+        jws.setAlgorithmHeaderValue(algorithm);
         try {
             return jws.getCompactSerialization();
         } catch (JoseException e) {